Fix crash when searching text
Oh my god. The assembly code for scan and iScan (edits.asm) expects an int-sized second parameter which measures the length of the text buffer. However, these functions were forward-declared with a short-sized parameter in the C++ code, which caused them to read the parameter with extra digits at the beginning, so they would keep searching for text matches beyond the buffer's limits. This was not an issue at compile time becase these functions use C linkage.

This is a bug in Turbo Vision's source code which only affects 32-bit builds. The binary library provided with PowerPack is exempt from this issue.

Tested against TVEDIT.

+++ source/tvision/teditor2.cpp
@@ -37,8 +37,8 @@
 
 extern "C" {
 int countLines( void *buf, uint count );
-ushort scan( const void *block, ushort size, const char *str );
-ushort iScan( const void *block, ushort size, const char *str );
+ushort scan( const void *block, uint size, const char *str );
+ushort iScan( const void *block, uint size, const char *str );
 };
 
 inline int isWordChar( int ch )
